# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Marcenaria API",
        "description": "Documentação da API da Marcenaria usando Laravel + Sanctum + Swagger",
        "contact": {
            "email": "suporte@marcenaria.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/testimonials": {
            "get": {
                "tags": [
                    "Testimonials"
                ],
                "summary": "Lista todos os testemunhos",
                "operationId": "90a96033ea15e645984db8f89b6c727d",
                "responses": {
                    "200": {
                        "description": "Lista de testemunhos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Testimonials"
                ],
                "summary": "Cria um novo testemunho",
                "operationId": "5a179b57e69eee5ce96406f17d491218",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "client_name"
                                ],
                                "properties": {
                                    "client_name": {
                                        "type": "string",
                                        "example": "Nome do cliente"
                                    },
                                    "localidade": {
                                        "type": "string",
                                        "example": "localidade ou cidade"
                                    },
                                    "text": {
                                        "type": "string",
                                        "example": "Conteúdo do testemunho"
                                    },
                                    "rating": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "photo_url": {
                                        "type": "string",
                                        "example": "Endereço da imagem"
                                    },
                                    "is_published": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Testemunho criado com sucesso"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/testimonials/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Exibe um testemunho específico",
                "operationId": "b97e9ea05c10c0033a762beef5a7cf2c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da testemunho",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Testemunho encontrada"
                    },
                    "404": {
                        "description": "Testemunho não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Testimonials"
                ],
                "summary": "Atualiza um testimunho",
                "operationId": "63f0799fd923400f9abcfe055746c19d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do testemunho",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "client_name": {
                                        "type": "string",
                                        "example": "Nome do cliente"
                                    },
                                    "localization": {
                                        "type": "string",
                                        "example": "Cidade ou localidade"
                                    },
                                    "text": {
                                        "type": "string",
                                        "example": "Resumo da experiência"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Testemunho atualizado com sucesso"
                    },
                    "403": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Testemunho não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Testimonials"
                ],
                "summary": "Exclui um testemunho",
                "operationId": "dd074e5d0be816d801485589244a8baa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do testemunho",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Testemunho excluído com sucesso"
                    },
                    "403": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Testemunho não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Realiza login e retorna token",
                "operationId": "c373bc4cffa4a56fc64e3979b8ee7b96",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "usuario@teste.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "senha123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado com sucesso, retorna token"
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Realiza logout e invalida o token",
                "operationId": "5ff2dc8b48bff52908d3821a0ecc9b3e",
                "responses": {
                    "200": {
                        "description": "Logout realizado com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use o token gerado no login como Bearer Token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}